test: share zalouser test helpers
