+++ extensions/zalouser/src/accounts.test-mocks.ts
@@ -1,0 +1,10 @@
+import { vi } from "vitest";
+import { createDefaultResolvedZalouserAccount } from "./test-helpers.js";
+
+vi.mock("./accounts.js", async (importOriginal) => {
+  const actual = (await importOriginal()) as Record<string, unknown>;
+  return {
+    ...actual,
+    resolveZalouserAccountSync: () => createDefaultResolvedZalouserAccount(),
+  };
+});
